Создал новый класс: BotOptions. Добавил новые константы: USERNAME, BASEURL, TOKEN. Переименовал класс Bot в TelegramBot.
